move streamdaq to devices
