+++ docs/api/devices/base.md
@@ -1,0 +1,7 @@
+# `devices.base`
+
+```{eval-rst}
+.. automodule:: mio.devices.base
+    :members:
+    :undoc-members:
+```
