+++ docs/api/devices/index.md
@@ -1,0 +1,8 @@
+# devices
+
+High-level control of external devices like miniscopes and DAQs
+
+```{toctree}
+base
+stream
+```
