+++ docs/cli/index.md
@@ -13,7 +13,7 @@ hash
 
 Refer to the following page for details regarding ``stream_daq`` device config files.
 
-- [stream_daq](../api/stream_daq.md)
+- [stream_daq](../api/devices/stream)
 
 ```{eval-rst}
 .. click:: mio.cli.main:cli
