+++ docs/index.md
@@ -24,14 +24,14 @@ device/update_controller
 :maxdepth: 1
 :caption: API:
 
-api/devices
+api/devices/index
 api/io
+api/interfaces
 api/logging
 api/models/index
 api/exceptions
 api/plots/index
 api/utils
-api/stream_daq
 api/bit_operation
 api/vendor/index
 api/process/index
