+++ mio/devices/__init__.py
@@ -1,0 +1,6 @@
+"""
+Main user-facing device control.
+"""
+
+from mio.devices.base import Device
+from mio.devices.stream import StreamDevice
