+++ mio/devices/base.py
@@ -1,0 +1,5 @@
+from abc import ABC, abstractmethod
+
+
+class Device(ABC):
+    """Abstract device parent class"""
