+++ tests/data/config/stream_daq_test_200px.yml
@@ -11,7 +11,7 @@ bitstream: "XEM7310-A75/USBInterface-8_33mhz-J2_2-3v3-IEEE.bit"
 # Preamble for each data buffer.
 preamble: 0x12345678
 
-# Image format. StreamDaq will calculate buffer size, etc. based on these parameters
+# Image format. StreamDevice will calculate buffer size, etc. based on these parameters
 frame_width: 200
 frame_height: 200
 pix_depth: 8
