+++ tests/data/config/wireless_example.yml
@@ -12,10 +12,10 @@ bitstream: "XEM7310-A75/USBInterface-8_33mhz-J2_2-3v3-IEEE.bit"
 port: null
 baudrate: null
 
-# Preamble for each data buffer. This is actually converted to bytes in the StreamDaq but has to be imported as a string because .yaml doesn't support hexadecimal formats.
+# Preamble for each data buffer. This is actually converted to bytes in the StreamDevice but has to be imported as a string because .yaml doesn't support hexadecimal formats.
 preamble: 0x12345678
 
-# Image format. StreamDaq will calculate buffer size, etc. based on these parameters
+# Image format. StreamDevice will calculate buffer size, etc. based on these parameters
 frame_width: 304
 frame_height: 304
 pix_depth: 8
